+++ test/unit/encryption.ts
@@ -21,7 +21,7 @@ describe("Asymmetric encryption", () => {
         "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."
     ]
     const buffers = messages.map(msg => Buffer.from(msg))
-    const randomMessages = (() => { const msgs = []; for (let i = 0; i < 80; i++) msgs.push(Wallet.createRandom().mnemonic); return msgs })()
+    const randomMessages = (() => { const msgs = []; for (let i = 0; i < 80; i++) msgs.push(Wallet.createRandom().mnemonic.phrase); return msgs })()
 
     it("Should encrypt and recover messages", () => {
         for (let i = 0; i < messages.length; i++) {